Move connector to separate class
